% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseFileSize.R
\name{databaseFileSize}
\alias{databaseFileSize}
\title{Pre-compute the size of binding sites of each bed file in database}
\usage{
databaseFileSize(database_dir = database_dir, output_path = output_path)
}
\arguments{
\item{database_dir}{directory path including the name of the database folder}

\item{output_path}{directory path of the output file, save the output dataframe to .csv file(naming: "database_dir"_bindingsize.csv) into the user provided "output_path"}
}
\description{
The function compute the size of binding sites of each bed file in database and save the output(dataframe containing two columns:filename and binding size) into a ../database_dir_bindingsize.csv file
}
\examples{
databaseFileSize(database_dir="/dir/database_dir", output_path="output_path" )
}
