% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardCalc.R
\name{jaccardCalc}
\alias{jaccardCalc}
\title{Calculates nucleotide overlap and total length}
\usage{
jaccardCalc(bed1, pwd)
}
\arguments{
\item{bed1}{The file path of a query bed file to be compared to the database files.}

\item{bed2}{The file path of a bed file in the database files to be compared to the query file.}
}
\value{
A vector of the nucleotide overlap and their respective nucleotide lengths. To be used as an internal function for PValue_nt and GPSMatch_nt.
}
\description{
The function calculates the nucleotide overlap between two bed files (query and hit) and their respective nucleotide lengths.
}
\examples{
jaccardCalc("/dir/bed1.txt","/dir/folder_dir")
}
