% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapCalc.R
\name{overlapCalc}
\alias{overlapCalc}
\title{Calculates nucleotide overlap only}
\usage{
overlapCalc(bed1, pwd)
}
\arguments{
\item{bed1}{The file path of a query bed file to be compared to the database files.}

\item{bed2}{The file path of a bed file in the database files to be compared to the query file.}
}
\value{
A value of the nucleotide overlap. To be used as an internal function for PValue_nt and GPSMatch_nt.
}
\description{
The function calculates the nucleotide overlap between two bed files (query and hit).
}
\examples{
overlapCalc("/dir/bed1.txt","/dir/bed2.txt")
}
