% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivPVal.R
\name{indivPVal}
\alias{indivPVal}
\title{Calculate significance of bed file similarity}
\usage{
indivPVal(n = 100, bed1, genome, bed2, output_path)
}
\arguments{
\item{n}{The number of background files generated in order to compute the p-value. As the n increases, the p-value will become more reliable, but the user should be aware that this will significantly increase the computing time. We have set a default n of 100.}

\item{bed1}{The file path of a query bed file to be compared to the database files.}

\item{genome}{The file path of a genome file, which should be tab delimited and structured as follows: <chromName><TAB><chromSize>. A pre-formatted hg19 genome file can be found on the Github.}

\item{bed2}{The file path of a bedfile in the database, or any bed file.}

\item{output_path}{The file path where the output .csv file will be generated}
}
\value{
A .csv file showing the p-value, pi-score, and a five number summary of the background file distribution compared to the original jaccard index.
}
\description{
The function compare two bed files (query and hit) to check if the observed degree of similarity between the query and the hit files may occur purely by chance.
}
\examples{
indivPVal(100,"/dir/bed1.txt","/dir/genome.txt","/dir/folder_dir","/dir/output_folder")
}
