% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSimilarity.R
\name{rankSimilarity}
\alias{rankSimilarity}
\title{Ranks the similarities between a given query bed file with multiple bed files (database files) based on the nucleotides length}
\usage{
rankSimilarity(
  queryBed = queryBed,
  database_dir = database_dir,
  output_path = output_path,
  output_frm_DatabaseFileSize = output_frm_DatabaseFileSize
)
}
\arguments{
\item{queryBed}{The file path of a query bed file to be compared to the database files.}

\item{database_dir}{The directory of a folder containing database files to be used for comparison with the query file.}

\item{output_path}{The file path where the output .csv file will be generated.}

\item{output_frm_DatabaseFileSize}{The file was generated from "DatabaseFileSize" function. Please also include the directory path of this file. The file contains the two columns: 1) BED file names in the database, and 2) pre-calculated length of binding sizes for each BED file}
}
\value{
A dataframe that shows the similarities of the query file to the database files ranked from greatest to least.
}
\description{
The function compares a given query bed file with multiple bed files (database files), and ranks the relative similarity between each file pairing, computed using jaccard indexes, where 0 has no similarities and 1 has an identical file. The function can also provide a significance value (p-value) of the similarity index based on user selection.
}
